Delete unused Nodes in a partition
Problem:
Nodes need to be deleted when they are no longer referenced (e.g. pool member
deleted).

Solution:
Added a method to the ServiceDeployer to determine which nodes are unreferenced and
delete them. Added unit test too.
